rename: config 폴더명, 파일명 변경
